Refactor: 스케줄 선택 기능 컴포넌트 분리 작업
