+++ app/(route)/activity/register/_components/file-input/index.tsx
